refactor: move survey analysis logic to processing service
